Replace use of Mix.env with test for config key
When using releases, the `Mix` module is not available and the app will crash when reading the `uri()` function.  This change modifies `uri()` to test for what it really cares about which is the existence of a config option called `:test_server_uri` which is dynamically built in the test setup.  Resolves #1.